geometry-service/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Geometry processing
ezdxf>=1.2.0
shapely>=2.0.0
networkx>=3.0
numpy>=1.24.0
scipy>=1.11.0

# PDF processing
pymupdf>=1.23.0
pdf2image>=1.16.0
Pillow>=10.0.0

# Vision AI
anthropic>=0.25.0
openai>=1.0.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Utils
pydantic>=2.0.0
python-dotenv>=1.0.0
